refactor(outline): lift shared Lunora planners into src/data/outline-plans
Converge spike plan* / seed / map-node on Dotflowy Node + tree.ts helpers.
Spike imports via @dotflowy/outline-plans Vite alias; bun + pnpm tests pass.
