+++ spikes/lunora-outline/src/outline/index.ts
@@ -1,35 +1,52 @@
-export type { OutlineNode, OutlinePlan, PlanPatch, NodeKind } from "./types.js";
-export { emptyPlan } from "./types.js";
-export { orderSiblings, chainDisagreements } from "./sibling-chain.js";
-export type { ChainDisagreement } from "./sibling-chain.js";
-export { buildTreeIndex, childrenOf, makeNode, parentKeyOf } from "./tree.js";
-export type { TreeIndex } from "./tree.js";
+/**
+ * Spike re-exports shared planners from repo `src/data/outline-plans/`.
+ * Bridge stays local (ADR 0004 handoff UI seam).
+ *
+ * Import path: Vite/vitest alias `@dotflowy/outline-plans` →
+ * `../../../src/data/outline-plans` (see vite.config.ts / vitest.config.ts).
+ */
+
+export type {
+  OutlineNode,
+  OutlinePlan,
+  PlanPatch,
+  NodeKind,
+  NodeDocLike,
+  SeedIfEmptyArgs,
+  SeedIfEmptyFn,
+  TreeIndex,
+  ChainDisagreement,
+} from "@dotflowy/outline-plans";
+
 export {
+  emptyPlan,
+  orderSiblings,
+  chainDisagreements,
+  buildTreeIndex,
+  childrenOf,
+  makeNode,
+  makeOutlineNode,
+  parentKeyOf,
   planInsertSibling,
   planIndent,
   planOutdent,
   planRemoveNode,
   planSetText,
   applyPlan,
-} from "./planners.js";
-export {
   rowToNode,
   docToNode,
   nodeToDocFields,
   nodeToRow,
   nodeToInsertFields,
-} from "./map-node.js";
-export type { NodeDocLike } from "./map-node.js";
-export {
-  rowsToOutlineNodes,
-  bridgeTreeIndex,
-  bridgeOrderedChildren,
-} from "./lunora-bridge.js";
-export {
   DEMO_SEED_TEXTS,
   DEMO_SEED_IDS,
   shouldSeedOutline,
   planSeedIfEmpty,
   seedEmptyOutline,
-} from "./seed.js";
-export type { SeedIfEmptyArgs, SeedIfEmptyFn } from "./seed.js";
+} from "@dotflowy/outline-plans";
+
+export {
+  rowsToOutlineNodes,
+  bridgeTreeIndex,
+  bridgeOrderedChildren,
+} from "./lunora-bridge.js";
