+++ src/data/outline-plans/index.ts
@@ -1,0 +1,36 @@
+export type { OutlineNode, OutlinePlan, PlanPatch, NodeKind } from "./types";
+export { emptyPlan } from "./types";
+
+export { orderSiblings, chainDisagreements } from "../sibling-chain";
+export type { ChainDisagreement } from "../sibling-chain";
+
+export { buildTreeIndex, childrenOf, makeNode, parentKeyOf } from "../tree";
+export type { TreeIndex } from "../tree";
+
+export {
+  makeOutlineNode,
+  planInsertSibling,
+  planIndent,
+  planOutdent,
+  planRemoveNode,
+  planSetText,
+  applyPlan,
+} from "./planners";
+
+export {
+  rowToNode,
+  docToNode,
+  nodeToDocFields,
+  nodeToRow,
+  nodeToInsertFields,
+} from "./map-node";
+export type { NodeDocLike } from "./map-node";
+
+export {
+  DEMO_SEED_TEXTS,
+  DEMO_SEED_IDS,
+  shouldSeedOutline,
+  planSeedIfEmpty,
+  seedEmptyOutline,
+} from "./seed";
+export type { SeedIfEmptyArgs, SeedIfEmptyFn } from "./seed";
